Add Changeset When Insert imp.

+++ src/Extensions/ModelBuilderExtensions.cs
@@ -20,7 +20,7 @@ namespace EntityFrameworkCore.AuditR.Extensions
                 b.Property(c => c.UserId).HasMaxLength(38).IsRequired();
                 b.Property(c => c.UserName).HasMaxLength(128).IsRequired();
                 b.Property(c => c.IPAddress).HasMaxLength(45).IsRequired();
-                b.Property(c => c.OperationType).HasColumnType("tinyint").IsRequired()
+                b.Property(c => c.OperationType).HasColumnType("int").IsRequired()
                 .HasConversion(v => (int)v, v => (OperationType)Enum.Parse(typeof(OperationType), v.ToString()));
             });
